ReplGPT.jl

Documentation for ReplGPT.jl

ReplGPT.jl

Key Management Functions

Conversation Management

Output Formatting

Index

Key Management Functions

ReplGPT.getAPIkey — Method

function getAPIkey()

Returns an OpenAI API key to use from either the LocalPreferences.toml file or the OPENAI_API_KEY environment variable. If neither is present, returns missing.

ReplGPT.setAPIkey — Method

function setAPIkey(key::String)

Sets the OpenAI API key for ReplGPT to use. The key will be saved as plaintext to your environment's LocalPreferences.toml file (perhaps somewhere like ~/.julia/environments/v1.8/LocalPreferences.toml). The key can be deleted with ReplGPT.clearAPIkeyI().

ReplGPT.clearAPIkey — Method

function clearAPIkey()

Deletes the OpenAI API key saved in LocalPreferences.toml if present.

See also: ReplGPT.setAPIkey(key::String)

Conversation Management

ReplGPT.initialize_conversation — Method

function ReplGPT.initialize_conversation()

Sets the ChatGPT conversation to an empty state. This effectively starts a new chat with ChatGPT with no recollection of past messages.

Missing docstring.

Missing docstring for ReplGPT.save_conversation(). Check Documenter's build log for details.

Output Formatting

ReplGPT.setFormatter — Method

function setFormatter(f::Function)

Set the ReplGPT formatter. The default formatter is ReplGPT.markdown, but ReplGPT.plaintext may be preferred if ChatGPT's response shouldn't be displayed as Markdown text.

Other functions may be passed to setFormatter as long as they accept a string as input an return or print something.

Examples

julia> ReplGPT.setFormatter(ReplGPT.plaintext)
plaintext (generic function with 1 method)

julia> ReplGPT.setFormatter(ReplGPT.markdown)
markdown (generic function with 1 method)

ReplGPT.markdown — Method

function markdown(s::String)

A simple Markdown formatter. This is the default since ChatGPT seems to like emitting Markdown strings and the formatting looks nice.

See also: ReplGPT.plaintext, another formatter that does no formatting, which is nice in case you don't want the ChatGPT output to be displayed as Markdown.

ReplGPT.plaintext — Method

function plaintext(s::String)

A simple plain text formatter. Unlike ReplGPT.markdown, this formatter performs no text transformations except removing some leading whitespace.

Index

ReplGPT.clearAPIkey

ReplGPT.getAPIkey

ReplGPT.initialize_conversation

ReplGPT.markdown

ReplGPT.plaintext

ReplGPT.setAPIkey

ReplGPT.setFormatter